Add loading weights from memory
Since one use case of the library is a WASM build we can't necessarily
rely on always having access to a file system. It makes sense to
introduce something that allows us to load model weights in the same
format as tensor_load_from_blob but from memory instead of a file
stream.